Add passcode length to view config, remove AppLockConigData from view controllers.  Move all related files under AppLock view folder.
