+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/Security/AppLockView/SFAppLockViewControllerTypes.h
@@ -1,8 +1,5 @@
 /*
- SFSDKPasscodeTextField.h
- SalesforceSDKCore
- 
- Copyright (c) 2018-present, salesforce.com, inc. All rights reserved.
+ Copyright (c) 2014-present, salesforce.com, inc. All rights reserved.
  
  Redistribution and use of this software in source and binary forms, with or without modification,
  are permitted provided that the following conditions are met:
@@ -25,36 +22,19 @@
  WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#import <UIKit/UIKit.h>
-#import "SFSDKAppLockViewConfig.h"
+#import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
-@protocol SFSDKPasscodeDeleteProtocol
-
--(void)deleteBackward;
-
-@end
-
-@interface SFSDKPasscodeTextField : UITextField
-
-@property (weak,nonatomic) id deleteDelegate;
-
-@property (nonatomic) BOOL passcodeLengthKnown;
-
-@property (assign,assign) NSUInteger passcodeLength;
-
-@property (strong,nonatomic) NSMutableString *passcodeInput;
-
-- (void)clearPasscode;
-
-- (void)refreshView;
-
-- (instancetype)initWithFrame:(CGRect)frame andLength:(NSUInteger)length;
-
-- (instancetype)initWithFrame:(CGRect)frame andLength:(NSUInteger)length andViewConfig:(SFSDKAppLockViewConfig *)config;
-
-@end
+/**
+ Mode constants indicating whether to create/verify an passcode or use biometric.
+ */
+typedef NS_ENUM(NSUInteger, SFAppLockControllerMode) {
+    SFAppLockControllerModeCreatePasscode,
+    SFAppLockControllerModeVerifyPasscode,
+    SFAppLockControllerModeChangePasscode,
+    SFAppLockControllerModeEnableBiometric,
+    SFAppLockControllerModeVerifyBiometric
+} NS_SWIFT_NAME(AppLockViewController.AppLockControllerMode);
 
 NS_ASSUME_NONNULL_END
-
