+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/Security/AppLockView/SFSDKPasscodeVerifyController.h
@@ -34,24 +34,16 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)passcodeVerified:(NSString *)passcode;
 
+- (void)passcodeFailed;
+
 @end
 
 @interface SFSDKPasscodeVerifyController : UIViewController
 
 /**
- * The configuration data used to verify the passcode.
- */
-@property (readonly) SFAppLockConfigurationData configData;
-
-/**
  Setup passcode view related preferences.
  */
 @property (nonatomic, readonly) SFSDKAppLockViewConfig *viewConfig;
-
-/**
- * Known length of the user's passcode.
- */
-@property (nonatomic) NSUInteger passcodeLength;
 
 /**
  * Wether the passcode length is known.
@@ -60,7 +52,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic,weak) id <SFSDKPasscodeVerifyDelegate> verifyDelegate;
 
-- (instancetype)initWithPasscodeConfigData:(SFAppLockConfigurationData)configData viewConfig:(SFSDKAppLockViewConfig *)config;
+- (instancetype)initWithViewConfig:(SFSDKAppLockViewConfig *)config;
 
 @end
