+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/Security/AppLockView/UIColor+SFSDKPasscodeView.m
